Intorduce new ClassInfo POSITIONAL_ARG_DEFAULTS
Argument defaults saved in this dict will be used, if no Value was provided by user and a TypeError occures when initialising an object.

E.g. idShort must be set to a value when initialising Referable object, but it is optional, so no widget will be created for it in AddDialog, only if user creates a widget for idShort, the corresponding value will be provided for init method, otherwise no value for idShort argument will be provided and an Exception will be raised