Initialize icons adter QApp initialization
There was a problem with pyinstaller, that
it returned error when trying to build the app.

The error was, when it tried to import icons.py.
It could not be imported as it used QPixmap inside,
which cannot be used without QApplication.
So we initialize icons after the start of the app.